refactor(mediator): step1: separate concerns,explicit pathname
