+++ mediator/src/bin/client-web.rs
@@ -1,6 +1,6 @@
 /// Client-side focused api accessible Aries Agent
 use log::info;
-use mediator::{agent::AgentMaker, routes::client::build_client_router};
+use mediator::{aries_agent::AgentMaker, http_routes::client::build_client_router};
 
 #[tokio::main]
 async fn main() {
