+++ mediator/src/http_routes/mod.rs
@@ -22,7 +22,7 @@ use xum_test_server::{
     storage::MediatorPersistence,
 };
 
-use crate::{agent::Agent, utils::string_from_std_error};
+use crate::{aries_agent::Agent, utils::string_from_std_error};
 type ArcAgent<T, P> = Arc<Agent<T, P>>;
 
 pub mod client;
